fix: resolve compilation errors and add macOS CI build
- Fix TokenType name conflict with Windows winnt.h (rename to PicoTokenType)
- Fix ObjMutex/ObjChannel typedef conflict in thread.h
- Add missing stdbool.h to lexer.h
- Move _POSIX_C_SOURCE before all includes in interpreter.c
- Add struct_def_new/instance_new declarations to value.h
- Add ObjThread full definition visibility in value.c
- Add string.h to thread.c
- Fix %lld format for int64_t in json.c
- Suppress unused variable warning in net.c
- Add macOS build job to release.yml with artifact upload

+++ src/stdlib/json.c
@@ -7,7 +7,7 @@
 static void stringify_value(Value v, char *buf, int *pos) {
     if (IS_NIL(v)) { *pos += sprintf(buf + *pos, "null"); }
     else if (IS_BOOL(v)) { *pos += sprintf(buf + *pos, v.boolean ? "true" : "false"); }
-    else if (IS_INT(v)) { *pos += sprintf(buf + *pos, "%lld", v.integer); }
+    else if (IS_INT(v)) { *pos += sprintf(buf + *pos, "%ld", (long)v.integer); }
     else if (IS_FLOAT(v)) { *pos += sprintf(buf + *pos, "%g", v.floating); }
     else if (IS_STR(v)) { *pos += sprintf(buf + *pos, "\"%s\"", v.string->data); }
     else if (IS_LIST(v)) {
